+++ Campaign/resources/assets/js/banner.js
@@ -166,6 +166,8 @@ window.remplib = window.remplib || {};
                     requestHeaders: model['newsletter_rectangle_template']['request_headers'] || null,
                     paramsTr: model['newsletter_rectangle_template']['params_tr'] || null,
                     paramsExtra: model['newsletter_rectangle_template']['params_extra'] || null,
+                    responseFailure: model['newsletter_rectangle_template']['response_failure'] || null,
+                    timeoutMessage: model['newsletter_rectangle_template']['timeout_message'] || null,
                 }
             }